+++ OM4labs/test/test_common.py
@@ -15,7 +15,7 @@ grid_1x1["lat_bnds"] = xr.DataArray(
 
 
 def test_infer_and_assign_coord():
-    from OM4_Analysis_Labs.om4common import infer_and_assign_coord
+    from OM4labs.om4common import infer_and_assign_coord
 
     da = xr.DataArray(data=np.empty((360, 2)), dims=("lon", "bnds"))
     da = infer_and_assign_coord(grid_1x1, da, "lon")
@@ -23,7 +23,7 @@ def test_infer_and_assign_coord():
 
 
 def test_copy_coordinates():
-    from OM4_Analysis_Labs.om4common import copy_coordinates
+    from OM4labs.om4common import copy_coordinates
 
     da = xr.DataArray(data=np.empty((360, 2)), dims=("lon", "bnds"))
     da = copy_coordinates(grid_1x1["lon_bnds"], da, ["lon", "bnds"])
@@ -31,7 +31,7 @@ def test_copy_coordinates():
 
 
 def test_compute_area_regular_grid():
-    from OM4_Analysis_Labs.om4common import compute_area_regular_grid
+    from OM4labs.om4common import compute_area_regular_grid
 
     area = compute_area_regular_grid(grid_1x1)
     assert isinstance(area, np.ndarray)
