+++ setup.py
@@ -1,15 +1,15 @@
-''' setup for OM4_Analysis_Labs '''
+''' setup for OM4labs '''
 import setuptools
 
 setuptools.setup(
-    name="OM4_Analysis_Labs",
+    name="OM4labs",
     version="0.0.1",
     author="Raphael Dussin",
     author_email="raphael.dussin@gmail.com",
     description=("A tool to analyze OM4 model output"),
     license="GPLv3",
     keywords="",
-    url="https://github.com/raphaeldussin/OM4_Analysis_Labs",
-    packages=['OM4_Analysis_Labs'],
-    scripts=['OM4_Analysis_Labs/annual_bias_1x1deg.py']
+    url="https://github.com/raphaeldussin/OM4labs",
+    packages=['OM4labs'],
+    scripts=['OM4labs/diags/annual_bias_1x1deg/annual_bias_1x1deg.py']
 )
